forgjprew miles long to int v1.1

+++ src/main/java/RewardValue.java
@@ -1,12 +1,12 @@
 public class RewardValue {
     public static final double CONVERSION_MILES_TO_RATE = 0.0035;
     double cash;
-    long miles;
+    int miles;
 
     public RewardValue(double cash) {
         this.cash = cash;
     }
-    public RewardValue(long miles)
+    public RewardValue(int miles)
     {
         this.miles = miles;
     }
@@ -15,9 +15,9 @@ public class RewardValue {
         this.cash = this.miles / RewardValue.CONVERSION_MILES_TO_RATE ;
         return this.cash;
     }
-    public long getMilesValue()
+    public int getMilesValue()
     {
-        this.miles = (long) (RewardValue.CONVERSION_MILES_TO_RATE * this.cash);
+        this.miles = (int) (RewardValue.CONVERSION_MILES_TO_RATE * this.cash);
         return this.miles;
     }
 }
